requirement.txt
Django
pillow
django-cleanup
django-allauth
django-htmx
python-dotenv
mysqlclient
shortuuid
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-prometheus